+++ galsim/angle.py
@@ -216,13 +216,13 @@ def hms(self, sep=":"):
     An optional `sep` parameter can change the : to something else (e.g. a space or
     nothing at all).
 
-    Note: the reverse process is effected by HMS_Angle:
+    Note: the reverse process is effected by Angle.from_hms:
 
         >>> angle = -5.357 * galsim.hours
         >>> hms = angle.hms()
         >>> print hms
         +18:38:34.80000000
-        >>> angle2 = galsim.HMS_Angle(hms)
+        >>> angle2 = galsim.Angle.from_hms(hms)
         >>> print angle2 / galsim.hours
         18.643
         >>> print angle2 / galsim.hours - 24
@@ -240,17 +240,17 @@ def hms(self, sep=":"):
     return _make_dms_string(h,sep)
 
 def dms(self, sep=":"):
-    """Return a DMS representation of the angle as a string: (+/-)ddmmss.decimal
+    """Return a DMS representation of the angle as a string: (+/-)dd:mm:ss.decimal
     An optional `sep` parameter can change the : to something else (e.g. a space or
     nothing at all).
 
-    Note: the reverse process is effected by DMS_Angle:
+    Note: the reverse process is effected by Angle.from_dms:
 
         >>> angle = -(5 * galsim.degrees + 13 * galsim.arcmin + 23 * galsim.arcsec)
         >>> dms = angle.dms()
         >>> print dms
         -05:13:23.00000000
-        >>> angle2 = galsim.DMS_Angle(dms)
+        >>> angle2 = galsim.Angle.from_dms(dms)
         >>> print angle2 / galsim.degrees
         -5.22305555556
         >>> print angle2 - angle
@@ -294,7 +294,7 @@ def parse_dms(s):
 
     return sign * (d + m/60. + s/3600.)
 
-def HMS_Angle(str):
+def Angle_from_hms(cls, hms_string):
     """Convert a string of the form hh:mm:ss.decimal into an Angle.
 
     There may be an initial + or - (or neither), then two digits for the hours, two for the
@@ -308,7 +308,7 @@ def HMS_Angle(str):
         >>> hms = angle.hms()
         >>> print hms
         +18:38:34.80000000
-        >>> angle2 = galsim.HMS_Angle(hms)
+        >>> angle2 = galsim.Angle.from_hms(hms)
         >>> print angle2 / galsim.hours
         18.643
         >>> print angle2 / galsim.hours - 24
@@ -318,9 +318,9 @@ def HMS_Angle(str):
 
     @returns the corresponding Angle instance
     """
-    return parse_dms(str) * galsim.hours
+    return parse_dms(hms_string) * galsim.hours
 
-def DMS_Angle(str):
+def Angle_from_dms(cls, dms_string):
     """Convert a string of the form dd:mm:ss.decimal into an Angle.
 
     There may be an initial + or - (or neither), then two digits for the degrees, two for the
@@ -330,7 +330,10 @@ def DMS_Angle(str):
 
     @returns the corresponding Angle instance
     """
-    return parse_dms(str) * galsim.degrees
+    return parse_dms(dms_string) * galsim.degrees
+
+Angle.from_hms = classmethod(Angle_from_hms)
+Angle.from_dms = classmethod(Angle_from_dms)
 
 set_func_doc(Angle.wrap, """Wrap Angle to lie in the range [-pi, pi) radians.